requirements.txt
pytest
Faker
pylint
python-dotenv
requests
